requirements.txt
# Core dependencies
pandas>=2.1.0
numpy>=1.26.0
polars[pyarrow]>=0.19.3
yfinance>=0.2.28
statsmodels>=0.14.1
seaborn>=0.13.0
pmdarima>=2.0.4
arch>=6.3.0
quantstats>=0.0.62
plotly>=5.18.0
scipy>=1.11.0

# Time series & forecasting
prophet>=1.1.5
tsfresh>=0.20.1
sktime>=0.21.0

# ML & Optimization
scikit-learn>=1.3.2
pyportfolioopt>=1.5.5
optuna>=3.3.0
xgboost>=2.0.3
lightgbm>=4.1.0
#catboost>=1.2.3

# Backtesting & trading
backtrader>=1.9.78.123
vectorbt>=0.25.4
bt>=0.2.8

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
pytorch-forecasting>=1.0.0

# Feature engineering & preprocessing
category-encoders>=2.6.0
feature-engine>=1.6.2

# Utils & IO
diskcache>=5.6.3
pyyaml>=6.0.1
python-dotenv>=1.0.1
pyarrow>=14.0.1
cython>=3.0.6
joblib>=1.3.2
tqdm>=4.66.1

# Monitoring & Experiment tracking
neptune-client>=1.3.0
wandb>=0.16.0
mlflow>=2.8.1

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
ipywidgets>=8.1.1

# Dev tools & Testing
pytest>=7.4.2
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.3

# Data versioning & pipelines
dvc>=3.41.0
pre-commit>=3.5.0

# API & Web integration
fastapi>=0.104.1
uvicorn>=0.24.0.post1
requests>=2.31.0

# Parallel processing
dask>=2023.11.0
#ray>=2.8.1

# Geospatial & Alternative data
geopandas>=0.14.0
ta>=0.11.0